ASR-012 remove approver main actor escape hatch
